requirements.txt
opik
pydantic
python-dotenv
google-generativeai
requests
python-dateutil
pytest
